[Feat] - skill type 파일 분리
